Allow types to declare their table name.
Some types may be direct descendants of another type, but have its
contents exist on a different table. For instance, there may be a table
of `items`, and one might want to be able to add many tags to those
items, so they might have a `item_tags` table associated with it. In Zig
code the most logical way to represent this would be a `Item` type, and
a `Tag` type name-spaced within it, like `Item.Tag`. The current code
would require two unrelated types, `Items` & `ItemTags`, especially if
there might be an `Item2` with its own associated tags.

This allows a user to optionally change the table name for a given type,
by setting a declaration:

    pub const sql_table_name = "item_tags";

Figured it made sense to give it a name that wasn't likely to be used
otherwise, although maybe `table_name` would have sufficed?